requirements.txt
Flask
Flask-Cors
gunicorn
firebase-admin
Pyrebase4